Move neovim plugin configurations into lua/config, add asciiart plugin
