+++ conf/nvim/lua/config/devicons.lua
