+++ conf/nvim/lua/config/navic.lua
