+++ conf/nvim/lua/config/nvim-tree.lua
