+++ conf/nvim/lua/config/rust-tools.lua
@@ -1,6 +1,6 @@
 local rt = require('rust-tools')
 
-local lsp = require('lsp')
+local lsp = require('config.lsp')
 local rust_analyzer_on_attach = function(client, bufnr)
   lsp.on_attach(client, bufnr)
   -- Hover actions
