+++ conf/nvim/lua/config/startify.lua
