+++ conf/nvim/lua/config/treesitter.lua
