+++ conf/nvim/lua/config/trouble.lua
